refactor: restructure library organization and examples for improved scalability
Implement comprehensive restructuring of the af_animations library to improve
code organization, maintainability, and developer experience as the widget
collection expands.

Structural changes:
- Reorganize lib/ directory with semantic folder structure:
  • property_animations/ (formerly value/) - property-based animations
  • layout_animations/ (formerly implicitly/) - layout transition animations
  • utility_widgets/ (formerly other/) - helper and utility components
- Restructure example/ directory for enhanced readability and comprehension
- Improve logical grouping of related animation components

Benefits:
- Enhanced discoverability of animation widgets by category
- Simplified navigation for developers exploring the library
- Scalable structure supporting future widget additions
- More intuitive organization aligned with animation use cases
- Clearer separation of concerns between different animation types

This reorganization establishes a solid foundation for continued library growth
while maintaining backward compatibility and improving the overall developer
experience.
