+++ example/lib/effect_animations_examples/effect_animations_page.dart
@@ -1,0 +1,100 @@
+import 'package:flutter/material.dart';
+
+
+class EffectAnimationsPage extends StatefulWidget {
+  const EffectAnimationsPage({
+    super.key,
+  });
+
+  @override
+  State<EffectAnimationsPage> createState() => _EffectAnimationsPageState();
+}
+
+class _EffectAnimationsPageState extends State<EffectAnimationsPage> {
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: PreferredSize(
+        preferredSize: Size(MediaQuery.of(context).size.width, kToolbarHeight),
+        child: AppBar(
+          title: const Text('Effect Animations'),
+          centerTitle: true,
+        ),
+      ),
+      body: SafeArea(
+        child: Center(
+          child: SingleChildScrollView(
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                const SizedBox(height: 30),
+        
+                SizedBox(
+                  width: MediaQuery.of(context).size.width * 0.6,
+                  child: const ElevatedButton(
+                    onPressed: null, ///In progress
+                    child: Text("AfFadeIn", textAlign: TextAlign.center)
+                  ),
+                ),
+
+                const SizedBox(height: 15),
+        
+                SizedBox(
+                  width: MediaQuery.of(context).size.width * 0.6,
+                  child: const ElevatedButton(
+                    onPressed: null, ///In progress
+                    child: Text("AfSlideInLeft", textAlign: TextAlign.center)
+                  ),
+                ),
+        
+                const SizedBox(height: 15),
+        
+                SizedBox(
+                  width: MediaQuery.of(context).size.width * 0.6,
+                  child: const ElevatedButton(
+                    onPressed: null, ///In progress
+                    child: Text("AfSlideInRight", textAlign: TextAlign.center)
+                  ),
+                ),
+        
+                const SizedBox(height: 15),
+        
+                SizedBox(
+                  width: MediaQuery.of(context).size.width * 0.6,
+                  child: const ElevatedButton(
+                    onPressed: null, ///In progress
+                    child: Text("AfSlideInUp", textAlign: TextAlign.center)
+                  ),
+                ),
+        
+                const SizedBox(height: 15),
+        
+                SizedBox(
+                  width: MediaQuery.of(context).size.width * 0.6,
+                  child: const ElevatedButton(
+                    onPressed: null, ///In progress
+                    child: Text("AfSlideInDown", textAlign: TextAlign.center)
+                  ),
+                ),
+        
+                const SizedBox(height: 15),
+        
+                SizedBox(
+                  width: MediaQuery.of(context).size.width * 0.6,
+                  child: const ElevatedButton(
+                    onPressed: null, ///In progress
+                    child: Text("AfScaleIn", textAlign: TextAlign.center)
+                  ),
+                ),
+        
+                const SizedBox(height: 30),
+              ],
+            ),
+          )
+        ),
+      ),
+    );
+  }
+
+}
