+++ example/lib/layout_animations_examples/animated_clip_rrect/animated_clip_rrect_page_optimized.dart
