+++ example/lib/property_animations_examples/animated_composed/animated_composed_page_optimized.dart
