+++ example/lib/property_animations_examples/animated_value/animated_value_page_optimized.dart
