+++ example/lib/utility_widgets_examples/utility_widgets_page.dart
@@ -1,0 +1,67 @@
+import 'package:af_animations_example/select_example_type.dart';
+import 'package:flutter/material.dart';
+
+
+class UtilityWidgetsPage extends StatefulWidget {
+  const UtilityWidgetsPage({
+    super.key,
+  });
+
+  @override
+  State<UtilityWidgetsPage> createState() => _UtilityWidgetsPageState();
+}
+
+class _UtilityWidgetsPageState extends State<UtilityWidgetsPage> {
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: PreferredSize(
+        preferredSize: Size(MediaQuery.of(context).size.width, kToolbarHeight),
+        child: AppBar(
+          title: const Text('Utility Widgets'),
+          centerTitle: true,
+        ),
+      ),
+      body: SafeArea(
+        child: Center(
+          child: SingleChildScrollView(
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                const SizedBox(height: 30),
+        
+                SizedBox(
+                  width: MediaQuery.of(context).size.width * 0.6,
+                  child: ElevatedButton(
+                    onPressed: () {
+                      Navigator.of(context).push(
+                        MaterialPageRoute(builder: (context) => SelectExampleTypePage(
+                          description: "The 'Builder' version, which has AfWidget implemented inside.",
+                          titleUnoptimizeVersion: "Without using AfBuilder",
+                          onTapUnoptimizeVersion: () {
+                            Navigator.of(context).pushNamed("/builder_page_unoptimized");
+                          },
+                          titleOptimizeVersion: "Using AfBuilder",
+                          onTapOptimizedVersion: () {
+                            Navigator.of(context).pushNamed("/builder_page_optimized");
+                          },
+                          title: "AfBuilder Demo"
+                        ))
+                      );
+                    },
+                    child: const Text("AfBuilder", textAlign: TextAlign.center)
+                  ),
+                ),
+        
+                const SizedBox(height: 30),
+              ],
+            ),
+          )
+        ),
+      ),
+    );
+  }
+
+}
+
