+++ lib/af_animations.dart
@@ -12,12 +12,12 @@ part 'af_widget.dart';
 part 'af_widget_on.dart';
 part 'af_widget_state.dart';
 part 'extensions/widgets_extension.dart';
-part 'implicitly/animated_clip_rrect.dart';
-part 'other/builder.dart';
-part 'value/animated_color.dart';
-part 'value/animated_composed.dart';
-part 'value/animated_secondary_value.dart';
-part 'value/animated_value.dart';
+part 'layout_animations/animated_clip_rrect.dart';
+part 'property_animations/animated_color.dart';
+part 'property_animations/animated_composed.dart';
+part 'property_animations/animated_secondary_value.dart';
+part 'property_animations/animated_value.dart';
+part 'utility_widgets/builder.dart';
 
 /// {@template AfAnimations}
 /// The AfAnimations widget serves to globally set the duration and curve for all used AfWidgets.
